+++ nextjs-blog/components/website/Footer.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Footer = () => {
+    return (
+        <footer>
+            <p>
+            Created By Carlo Staltari, Mohaned Bouzaidi & Yan Burton
+            <br />
+            Champlain College ECP Final Project 2022-2023
+            </p>
+        </footer>
+    )
+}
+
+export default Footer;
