+++ nextjs-blog/pages/visit/index.js
@@ -2,6 +2,8 @@ import Head from "next/head";
 import styles from "../../styles/Home.module.css";
 import { useRouter } from "next/router";
 import VisitTable from "../../components/visit/VisitTable";
+import VisitCardView from "../../components/visit/VisitCardView";
+import Footer from "../../components/website/Footer";
 
 
 export async function getServerSideProps() {
@@ -14,27 +16,6 @@ export async function getServerSideProps() {
 }
 
 export default function Home({visits}) {
-  const router = useRouter()
-  function formatEmployees(emp_ids){
-    var formattedEmployees = ""
-    const employees = emp_ids.map((emp) =>
-      formattedEmployees = formattedEmployees + " " + emp
-    );
-    return formattedEmployees;
-  }
-  function formatDate(date){
-    let newDate = new Date(date);
-    let dd = newDate.getDate()+1;
-    let mm = newDate.getMonth()+1;
-    const yyyy = newDate.getFullYear();
-
-    if (dd < 10) dd = '0' + dd;
-    if (mm < 10) mm = '0' + mm;
-    const formattedDate = yyyy + '-' + mm + '-' + dd;
-    // console.log(formattedDate);
-
-    return formattedDate;
-  }
   return (
     <div className={styles.container}>
       <Head>
@@ -62,42 +43,7 @@ export default function Home({visits}) {
       </header>
       <main>
         <h2 className={styles.title}>All Visits</h2>
-        <div id="visitContainer">
-          {/* {visits.map((visit) => (
-            <div className="visit">
-              <div className="info">
-                <div className="vrRow">
-                  Visit:{visit.visit_id} Project: {visit.project_id}
-                </div>
-                <div className="empRow">
-                  Employee(s): {formatEmployees(visit.employee_ids)}
-                </div>
-                <div className="startRow">
-                  Start Date: {formatDate(visit.start_date)}
-                </div>
-                <div className="endRow">
-                  End Date: {formatDate(visit.end_date)}
-                </div>
-              </div>
-              <div className="editBtnDiv">
-                <button
-                  className="editBtn"
-                  name="edit"
-                  value={visit.visit_id}
-                  onClick={() =>
-                    router.push({
-                      pathname: "/visit/[id]",
-                      query: { id: visit.visit_id },
-                    })
-                  }
-                >
-                  Edit
-                </button>
-              </div>
-            </div>
-          ))} */}
-        </div>
-
+        {/* <VisitCardView visits={visits}/> */}
         <VisitTable visits={visits}/>
 
         <div id="root"></div>
@@ -116,13 +62,7 @@ export default function Home({visits}) {
         </div>
       </main>
 
-      <footer>
-        <p>
-          Created By Carlo Staltari, Mohaned Bouzaidi & Yan Burton
-          <br />
-          Champlain College ECP Final Project 2022-2023
-        </p>
-      </footer>
+      <Footer/>
 
       <style jsx>{`
         header {
