+++ nextjs-blog/styles/globals.css
@@ -36,3 +36,24 @@ button {
   padding: 0.5rem 1rem;
   font-weight: bold;
 }
+
+footer {
+  width: 100vw;
+  height: 100px;
+  border-top: 1px solid #eaeaea;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: #222222;
+  color: #ffffff;
+}
+footer img {
+  margin-left: 0.5rem;
+}
+footer p {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-decoration: none;
+  color: inherit;
+}
